bg-gradient, fall-color, url query, language render

+++ dragblock/dragblock/build/applications/patterns/server/patterns-loader.php
@@ -8,25 +8,29 @@
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
-// dev-reply#52.
-remove_theme_support( 'core-block-patterns' );
-// dev-reply#55.
+// dev-reply#93.
 add_action( 'rest_api_init', 'dragblock_patterns_rest_api_init', 1 );
-// dev-reply#57.
+// dev-reply#935.
 /**
- * Check Documentation#57
+ * Check Documentation#95
  */
 function dragblock_patterns_rest_api_init() {
-	$dragblock_pl_pattern = 60 * 60 * 24; // dev-reply#511.
+	// dev-reply#938.
+	$dragblock_pl_debug = false;
+	$dragblock_pl_pattern = 60 * 60 * 24; // dev-reply#942.
 	$dragblock_pl_duration = DRAGBLOCK_K_PATTERN_CACHE . '--time';
-	delete_transient( $dragblock_pl_duration );
+	// dev-reply#945.
+	if ( $dragblock_pl_debug ) {
+		delete_transient( DRAGBLOCK_K_PATTERN_CACHE );
+		delete_transient( $dragblock_pl_duration );
+		// dev-reply#950.
+	}
 	$dragblock_pl_timer = get_transient( DRAGBLOCK_K_PATTERN_CACHE );
 	$dragblock_pl_key = get_transient( $dragblock_pl_duration );
-	// dev-reply#516.
 	$dragblock_pl_data = 'https://sneeit.com/api/validation/';
-	// dev-reply#519.
+	// dev-reply#960.
 	if ( ! $dragblock_pl_key || DRAGBLOCK_IS_LOCAL ) {
-		// dev-reply#524.
+		// dev-reply#965.
 		$dragblock_pl_time = wp_remote_request( $dragblock_pl_data, array(
 			'body' => array(
 				'get' => 'patterns',
@@ -35,33 +39,60 @@ function dragblock_patterns_rest_api_init() {
 		) );
 		if ( ! is_wp_error( $dragblock_pl_time ) ) {
 			$dragblock_pl_timer = json_decode( wp_remote_retrieve_body( $dragblock_pl_time ), true );
-			// dev-reply#535.
+			if ( $dragblock_pl_debug ) {
+				// dev-reply#976.
+			}
 			if ( ! empty( $dragblock_pl_timer['error'] ) || empty( $dragblock_pl_timer['patterns'] ) ) {
-				// dev-reply#538.
+				// dev-reply#980.
 				$dragblock_pl_timer = get_transient( DRAGBLOCK_K_PATTERN_CACHE );
 			} else {
-				// dev-reply#541.
+				// dev-reply#983.
 				set_transient( DRAGBLOCK_K_PATTERN_CACHE, $dragblock_pl_timer );
 				set_transient( $dragblock_pl_duration, true, $dragblock_pl_pattern );
 			}
 		} else {
-			// dev-reply#547.
+			if ( $dragblock_pl_debug ) {
+				// dev-reply#990.
+			}
 		}
+	}
+	if ( $dragblock_pl_debug ) {
+		// dev-reply#997.
 	}
 	if ( ! $dragblock_pl_timer ) {
-		// dev-reply#555.
+		// dev-reply#9103.
 		return;
 	}
+	// dev-reply#9109.
 	if ( ! empty( $dragblock_pl_timer['categories'] ) ) {
 		foreach ( $dragblock_pl_timer['categories'] as $dragblock_pl_api => $dragblock_pl_url ) {
+			$dragblock_pl_api = trim( explode( '-', $dragblock_pl_api )[0] );
+			if ( $dragblock_pl_debug ) {
+				// dev-reply#9115.
+				var_dump( $dragblock_pl_api );
+			}
 			register_block_pattern_category( $dragblock_pl_api, $dragblock_pl_url );
 		}
+		if ( $dragblock_pl_debug ) {
+			// dev-reply#9127.
+		}
 	}
-	// dev-reply#568.
 	foreach ( $dragblock_pl_timer['patterns'] as $dragblock_pl_response => $dragblock_pl_cat ) {
-		if ( empty( $dragblock_pl_cat['name'] ) ) {
+		if ( empty( $dragblock_pl_cat['name'] ) || empty( $dragblock_pl_cat['categories'] ) ) {
 			continue;
 		}
+		for ( $dragblock_pl_slug = 0; $dragblock_pl_slug < count( $dragblock_pl_cat['categories'] ); $dragblock_pl_slug++ ) {
+			$dragblock_pl_cat['categories'][ $dragblock_pl_slug ] = trim( explode( '-', $dragblock_pl_cat['categories'][ $dragblock_pl_slug ] )[0] );
+		}
+		if ( $dragblock_pl_debug ) {
+			var_dump( $dragblock_pl_cat );
+		}
+		if ( strpos( $dragblock_pl_cat['name'], 'dragblock/' ) === false ) {
+			$dragblock_pl_cat['name'] = 'dragblock/' . $dragblock_pl_cat['name'];
+		}
 		register_block_pattern( $dragblock_pl_cat['name'], $dragblock_pl_cat );
+	}
+	if ( $dragblock_pl_debug ) {
+		die();
 	}
 }
